Deletes temp directory before beginning download task
If the orchard download task fails during the extraction phase then
there is a good chance the `temp` directory will exist with files.
If the task is attempted again (which is most likely) it will likely
fail due to temp already existing. This would require the temp directory
to be manually deleted from the system. This change however ensures the
`temp` directory is deleted before the download task is executed.