Add remove modules task
Add a new task that allows the build script to remove modules that aren't
required from the build process. This will allow you to automatically
reduce the build time and deployment footprint of your Orchard install.
Simply pass the task an orchard install directory and a list of module
names (e.g.: Orchard.Taxonomies) and it will remove them from the SLN and
the local directory automatically.