Fix issue with extracting 1.9.2
There was an issue with extracting the latest release of Orchard (1.9.2).
The extraction was throwing an `EEXISTS` error complaining about a directory
in the `Orchard.AzureMediaServices` module. After invesitgation, it turns out
the `Scripts` directory in the module contains `Libs` & `libs` directories
which was causing an issue when creating the directory in the file system.
The solution was to add a check to see if the directory exists before creating
it.

Here is a link to the affected module on Github, showing the `Libs` & `libs`
directories.

https://github.com/OrchardCMS/Orchard/tree/1.9.2/src/Orchard.Web/Modules/Orchard.Azure.MediaServices/Scripts